tests/xtimer_cli: rework python integration
